+++ R/autoplot.R
@@ -3,8 +3,7 @@
 #' @param object  A matrix or a list of square matrices, with row and column names
 #'   indicating the age groups.
 #' @param ...	 Other arguments passed on
-#' @param title Title to give to plot setting matrices. Default title for plotting a single contact matrix is "Contact Matrices" and for
-#' setting wise contact matrix it is "Setting-specific synthetic contact matrices".
+#' @param title Title to give to plot setting matrices. Defaults are provided for certain objects
 #' @return a ggplot visualisation of contact rates
 #' @importFrom ggplot2 autoplot
 #' @name autoplot-conmat
@@ -47,8 +46,8 @@
 #' }
 #' @export
 autoplot.conmat_age_matrix <- function(object,
-                                              ...,
-                                              title = "Contact Matrices") {
+                                       ...,
+                                       title = "Contact Matrices") {
   plot_matrix(object) +
     ggplot2::ggtitle(title)
 }
@@ -58,7 +57,50 @@ autoplot.conmat_age_matrix <- function(object,
 autoplot.conmat_setting_prediction_matrix <- function(object,
                                                       ...,
                                                       title = "Setting-specific synthetic contact matrices") {
-  plot_setting_matrices(object,
+  plot_setting_matrices(
+    object,
     title = title
   )
+}
+
+#' @rdname autoplot-conmat
+#' @export
+autoplot.transmission_probability_matrix <- function(object,
+                                                     ...,
+                                                     title = "Setting-specific transmission probability matrices") {
+  plot_setting_matrices(
+    object,
+    title = title
+  ) +
+    ggplot2::labs(
+      subtitle = "Relative probability of individuals in an age group infecting an individual in another age group"
+    )
+}
+
+#' @rdname autoplot-conmat
+#' @export
+autoplot.ngm_setting_matrix <- function(object,
+                                        ...,
+                                        title = "Setting-specific NGM matrices") {
+  plot_setting_matrices(
+    object,
+    title = title
+  ) +
+    ggplot2::labs(
+      subtitle = "The number of newly infected individuals for a specified age group in each setting"
+    )
+}
+
+#' @rdname autoplot-conmat
+#' @export
+autoplot.setting_vaccination_matrix <- function(object,
+                                                ...,
+                                                title = "Setting-specific vaccination matrices") {
+  plot_setting_matrices(
+    object,
+    title = title
+  ) +
+    ggplot2::labs(
+      subtitle = "Number of newly infected individuals for age groups, adjusted based on proposed age group vaccination rates"
+    )
 }